Refactor into separate file
